fix: skip validation of `START_EPOCH` env variable
As the Altair epoch is different in different testnets, it is not
possible to set a static pre-defined number of the `START_EPOCH` as a
required minimum value of this variable. So, for now, we skip the
validation of this env variable on testnets.

Later it would be better to implement a custom validation directive for
validating the minimum possible value of the `START_EPOCH` variable
depending on the testnet where the application is running.