chore: add default port values to `docker-compose`
As stated in Readme, `HTTP_PORT` environment variable should be optional
and have a reasonable default value. But in the previous version of
`docker-compose` file it was assumed that the `HTTP_PORT` value must be
set. Now this assumption is eliminated and default values for
`HTTP_PORT` and `EXTERNAL_HTTP_PORT` environment variables are set in
`docker-compose.yml`.