fix: add default env values in docker-compose
As stated in `README.md`, some environmental variable values are
optional and so can be omitted in the `.env` config file. But these
variables didn't have default values in the `docker-compose.yml` config
file, so, it was not possible to omit these values when the app was
built using Docker. Now default values are added for all optional env
variables in the `docker-compose.yml`.