fix: simplify `DENCUN_FORK_EPOCH` validation
Simplify validation of the `DENCUN_FORK_EPOCH` environment variable. Now
the validation is based on these two assumptions:

1. If the value of the `DENCUN_FORK_EPOCH` variable is not provided, it
is now implicitly set to the maximum integer for any unknown networks.
So, the app assumes that by default the Dencun fork is not applied to
any unknown network.

2. For the 3 officially supported networks correct values for the
`DENCUN_FORK_EPOCH` variable are set implicitly. For these 3 networks
values that the user sets for the `DENCUN_FORK_EPOCH` variable are
silently ignored.

Custom validation directive and complex transform function have been
removed.

The app now logs the actual value of the `DENCUN_FORK_EPOCH` variable
when it starts.