feat: add dashboards with Simple DVT clusters info
Add a new dashboard that displays all registered operators (both
operators in the Curated module and operators in Simple DVT clusters).
For each operator, the new dashboard displays a list of Simple DVT
clusters in which this operator participates and the link to the
appropriate dashboard for this cluster.

Add a new dashboard that extends the existing "Node Operators"
dashboard. This new dashboard has a new dropdown panel with information
about simple DVT clusters. For operators in the Curated module, this
dropdown displays a list of Simple DVT clusters where this operator
participates. For Simple DVT clusters, the new dropdown panel displays
the list of participants of this cluster.

Information about Simple DVT clusters for new dashboards is obtained
from the special JSON file. The new "JSON API" data source has been
added to Grafana configs to fetch information for new dashboards from
the JSON file. Parameters of this data source can be specified via new
.env variables.

All new dashboards and data sources are optional. If none of the new
.env variables are set, the app will behave as previously.