feat: Grafana config variables validation
Implement Grafana env variables validation in the primary EVM app.

For the correct work of new dashboards new environmental variables must
always have valid values. To keep the previous dashboards untouched and
keep the previous working mode of the application, all new configuration
variables must be omitted in the .env config or have empty values. To
make the new dashboards work correctly, the new configuration variables
must have the correct values. If the value of the
`SIMPLE_DVT_MAPPING_FILE_PATH` is set up, it is assumed that the new
dashboards with Simple DVT clusters info should be enabled, and so, all
other variables also must have valid non-empty values.

As it is not possible to validate env variables inside the Grafana
container or validate them outside the app, it has been decided to pass
these variables inside the main EVM app only to check that the values of
these variables were set up correctly.

It is important to note that the EVM app doesn't use values of the new
variables. The only purpose of passing them inside the app container is
to implement validation of these variables. New variables should be
passed inside the container untouched to check their actual values in
the .env config file.