fix: links in Operators table
Previously for Simple DVT clusters on the "Node Operators" dashboard, if
an operator displayed in the "Operators" table was not in the Curated
module, it had a link that pointed out to the "Node Operators"
dashboard, which didn't make sense because the "Node Operators"
dashboard doesn't have a view for such operators.

Now operators in the "Operators" table that are not in the Curated
module have links that point out to the new "Operators List" dashboard.
All other operators in this table have links that point out to the "Node
Operators" dashboard.

Technically it is implemented this way:

1. Add two more fields to the JSON API query. These queries return the
dedicated dashboard name and ID for the particular operator in the
table.

2. Add two new hidden columns to the Operators table that correspond to
the abovementioned requests.

3. Use information from these two hidden fields to build the appropriate
link for the particular operator in the Operators table.