chore: add default DB config values into docker-compose
Previously when a user ran the application in a container and specified
incorrect database configuration values in the .env file, the
application failed to start. So, a user always must specify the correct
default Clickhouse config in .env variables. It was not very convenient.
A user actually should repeat the config specified in the
`docker-compose` or know the default database config values.

Now if the app runs in containers, clickhouse credentials, DB name, host
and port specified in the env variables are ignored and the default
values specified in the `docker-compose` are used.