fix: convert numeric labels to strings
1. Alertmanager failed to handle the numeric `nos_module_id` label
correctly. Because of this fact, Discord alerts with this label were not
sent and the app returned the Alertmanager error. Now this bug is fixed.

2. The app failed to work correctly with the latest version of the
Alertmanager. Now the particular version of the Alertmanager is fixed in
the `docker-compose` file so that the app can always start all its
containers correctly.

3. Increase memory required for the Prometheus service in the
docker-compose, so that the Prometheus service can correctly handle a
larger number of node operators.

4. Remove the `DISCORD_WEBHOOK_URL` env variable from env examples as we
don't have such a variable in the EVM app.

5. Slightly harmonize the content of alert labels.