fix: skip attestations from slots in previous epochs
As attestations for slots in all epochs except the current epoch and the
previous epoch are already handled during previous app cycles, we now
exclude these attestations from processing. So, the logic of attestation
processing is kept the same as it was before.

Also, update the `@lodestar/types` lib to support Pectra hardfork types.