feat: add compatibility with pre-Pectra forks
Add backward compatibility of the attestation calculation algorithm with
pre-Pectra forks. So, the app should work normally on the chain before
Pectra and after Pectra without any changes in app code or in configs.

Now information about fork epochs is fetched from the CL node instead of
setting it in the env variables or hard-coding these epochs for known
chains.