fix: link fork epoch processing to attestation
As the `AttestationService` fetches attestation for the epoch that is
processed on the current EVM cycle and for the epoch that was processed
on the previous cycle, an epoch in the `this.processedEpoch` does not
correctly reflect the epoch of the attestation that is currently
processed. So, the logic of how fork is determined worked incorrectly
for some attestations.

Now this bug is fixed and the fork is now determined using the slot of
each particular attestation.