fix: sync committee accounting
There was a bug in the accounting of sync committees. One validator
could participate in the same sync committee multiple times under
different committee indexes. Previously in the code, there was an
implicit assumption that a validator can be presented in a sync
committee of the particular block only one time. This led to an
incorrect accounting of sync committees and incorrect calculation of
validator rewards for participation in sync committees.

For example, in the 5-th Pectra devnet, the validator 8568
https://dora.pectra-devnet-5.ethpandaops.io/validator/0x9617f22db51844882a8115ee19613cf7a6211107a878aefc61111ad650304f0557aa10f977c38b9b978349c5eaf8ecc2
proposed a slot 179518
https://dora.pectra-devnet-5.ethpandaops.io/slot/179518
in the epoch 5609
https://dora.pectra-devnet-5.ethpandaops.io/epoch/5609
The sync committee for epoch 5609 contains validators with indexes 1298,
2561, 4326, 8113, and 9377 two times. Rewards for participation of these
validators in the sync committee for the second time were not taken into
account. So, when slot 179518 is proposed, EVM calculated rewards for
proposing this slot incorrectly.

Now the bug is fixed. The `sync_meta` field of the
`ValidatorDutySummary` structure was extended to contain several synced
blocks if a validator participate in the same committee multiple times.
Also, the calculation of the percentage of sync committee participation
and the calculation of rewards for sync committee participation have
been updated.