feat: configurable max slot deep count
Add a new `CL_API_MAX_SLOT_DEEP_COUNT` env variable that specifies a
maximum number of slots that the app uses to find a not missed
consensus-layer slot.

Previously, this number was hard-coded in the appropriate constant in
the `ConsensusProviderService`. As after the Pectra hardfork on Holeksy
blocks became proposed significantly more rarely than usual, it became
possible not to have any proposed blocks in the closest 32 slots, and
so, the app gave up trying to find the closed proposed block and threw
the error.

Now it is possible to configure this depth for cases like we see now on
Holesky and extend this depth to any necessary value. So the app can
look deeper into the slots chain and find the next (or previous)
proposed block.

Also, some minor refactoring has been done in the
`ConsensusProviderService` and fixed incorrect numbers in the error
message that is shown when the app cannot find the closest proposed
block. Previously in case of such errors, the error message always
looked like "Error when trying to get next not missed block header. From
N to N" because of the recursive call.