fix: better warnings in logs
When `LOG_FORMAT` is set to "simple", the `logger.warn(error)` command
logs only the non-default keys of the error object, but not the error
message itself. So, in the "simple" log format it is hard to notice an
important warning log, and merely impossible to determine the root cause
of the error by looking only at the log message.

Now when the app needs to log an error object, the log with the
"warning" level always logs the error message of this error object.