fix: error on writing to closed stream
When processing post-Pectra epochs on Hoodi, an error with writing data
to the DB has been discovered. If data that the app tries to write to
the Clickhouse table has an incorrect format, the app silently ignores
this error, doesn't write anything to the affected table, and proceeds
as if everything was OK.

This error happened because the procedure of writing data to the table
was wrapped with a retrier function. If an attempt to write data to the
DB is not successful, the retrier function tries to repeat the same
attempt several times more. But in some cases (for example, if data that
the app tries to save in the DB has an incorrect format) if the first
attempt was not successful when the retrier tries to repeat the failed
procedure again after some delay, the writing stream is already closed
and so, it is not possible to write anything into it. For some reason,
if the writing stream is closed, the attempt to write something into it
doesn't fail, but this attempt is skipped silently without any error.

Now the app explicitly checks the status of the stream before writing
anything into it. If a stream is closed, the app returns an error and
doesn't allow the code to proceed further.

Also, some unused methods and constants of the `Clickhouse` service are
removed in this PR.