fix: missed rewards for attestations
The current calculation algorithm of missed rewards for attestations is
working incorrectly for validators with a max effective balance greater
than 32. As a result of this calculation, the value of the missed
attestation reward for such validators might be negative. The attempt to
store a negative value of missed attestation reward to the DB causes
failures that lead to a complete inability to update the
"validators_summary" table and loss of data about all validators in the
epoch where this calculation error happened.

The calculation of missed attestation rewards for 0x02 validators is
temporarily disabled until we understand better how to calculate missed
attestation rewards for all types of validators.