fix: missed rewards for attestations - 2
Proper fix for missed attestations rewards calculation.

In the previous version of the missed rewards calculation algorithm,
there was a hard assumption that the effective balance of any validator
must be 32 ETH. After Pectra hardfork this assumption is no longer
valid. So, now to calculate the "perfect" attestation rewards that a
validator could earn if it fully executes all attestation duties the
algorithm should use the actual validator effective balance, not 32.

This PR makes the rewards calculation algorithm use the actual validator
balance in missed rewards calculation and also extracts some common
multiplications to new variables.