refactor: move common types to separate file
Move common general types (like `Epoch`, `Slot`, etc.) to a new
dedicated common `types.ts` file out of the `ConsensusProvider` module.
The presence of these types in the `ConsensusProvider` module is
problematic because it makes it more complex to add dependencies to the
`ConsensusProvider` module. As these common types are imported into many
other modules of the project, if someone adds a new dependency to the
`ConsensusProvider` module, there is a high chance that it will create a
cyclic dependency with the imported module. For example, if the
`Clickhouse` service is imported to the `ConsensusProvider` module, it
creates a cyclic dependency between the `ConsensusProvider` and
`Clickhouse` modules, because the `Clickhouse` module actively uses
general common types determined in the `ConsensusProvider` module.