feat: new algorithm for sparse networks
Implement a new algorithm for searching not missed CL slots. The new
algorithm should work much better for networks with a high number of
missed slots.

In the default mode, to find the next (or previous) not missed CL slot
the app iterates all next (or previous) CL slots one by one from the
currently processed slot to find the not missing one up to the depth
specified in the `CL_API_MAX_SLOT_DEEP_COUNT` variable.

If the new `SPARSE_NETWORK_MODE` is enabled, to find the next not
missing CL slot, the app gets the number of the EL block next to the EL
block corresponding to the latest known not missing slot, gets the
timestamp of this block, and gets the CL slot that corresponds to this
timestamp.

Also, add a new `last_not_missed_slot` column to the `epochs_metadata`
DB table. Now for each processed epoch, EVM saves to this new column
also the latest not missed slot for this epoch. Typically (if the
network conditions are normal) this slot is the latest (or one of the
latest) slot in the processed epoch. But if the network has very many
numbers of missed slots (even entire epochs that consist of missed
slots), the slot that is saved to the `last_not_missed_slot` column for
the processed epoch might be in some previous epochs. But this slot is
always the latest not missed slot for the processed epoch.

The pair `(epoch, last_not_missed_slot)` stored in the DB is used by EVM
in the new `SPARSE_NETWORK_MODE` to find necessary not-missed slots when
EVM processes data for `epoch + 1`.