fix: exclude pending initialized validators
Exclude validators with the "pending_initialized" status from the reward
calculation.

There was an error with the reward calculation for some validators on
Hoodi testnet on the epoch when the Pectra hardfork was activated
(2048).

At the time of the hadrfork, many validators of the "Attestant (BVI)
Limited" node operator were waiting for the activation. Most of these
validators had the status "pending_queued" at the last epoch before the
hardfork (2047). All validators with this status had 32 ETH on their
balance. On the epoch 2048 (Pectra hardfork activation epoch), some of
these validators changed their status from "pending_queued" to
"pending_initialized" (for example, validators with indexes from 1073611
to 1074310) and their balance became zero.

To calculate reward calculation error EVM compares all the rewards
calculated by EVM algorithms with the validators' balance change. As in
the previous epoch balance of many validators was 32, and in the next
epoch it became zero, it led to an incorrect reward calculation error.

To avoid this kind of issue in the future, validators with the
"pending_initialized" status are now excluded from the rewards
calculation. As these validators are not active anyway and don't
participate in validator duties, no rewards or penalties can be applied
to them.