refactor: optimize sparse network processing
Remove unnecessary checks of slot correctness. It turns out that the
timestamp of the block that the sparse network processing algorithm gets
from the next EL block is always accurate and always precisely
corresponds to the timestamp of the relevant CL slot.

Also, fill the block cache with the info about missing slots between the
neighboring proposed slots.

Add debug log record when the `getBlockHeader` or `getBlockInfo` methods
of the `ConsensusProvider` service cannot get the info from the block
cache. It can help to figure out how often the EVM app fetches data from
the CL node and in what cases.