fix: e2e tests
Fix broken end-to-end tests. As now the `ExecutionProviderService` is
required for the `ConsensusProviderService`, the `EL_RPC_URLS` env
variable must be set. Assign a fake value to this variable in jest tests
as the existing e2e tests don't test a new `SPARSE_NETWORK_MODE` and so,
don't actually use the EL RPC provider.