fix: getting attestation committee info
If the target epoch is too far from the target slot, the CL endpoint
`eth/v1/beacon/states/${slot}/committees?epoch=${epoch}`
fails to fetch data for committees in this slot. CL node returns an
error in this case.

Now if EVM sees that the target epoch is 2 epochs ahead of the target
slot's epoch, it assumes that the entire epoch after the target slot's
epoch is empty and so, the endpoint can safely fetch attestation
committee data for the epoch that is 1 epoch ahead from the target
slot's epoch. There should be no difference in attestation committee
data of the desired slot for this epoch and the actually requested
epoch.