feat: add validator IDs to alert metrics
Add up to 20 validator IDs to metrics corresponding to EVM alerts. Wrap
these validator IDs with a link to CL explorer set in the .env
variables. This allows the application to display in the Discord alerts
not only the number of validators affected by the particular event but
also some of these validators.