feat: support end of time range from the future on "Node Operators" Grafana dashboard
Adjust "Node Operators" Grafana dashboard variables to make the
dashboard work correctly if a user selects the end of the time range
from the future. In this case, the logic of the dashboard assumes that
the data for the last available epoch should be used.

Adjust time ranges in alerts, make the end value of the time range for
some alerts several hours after the alerting event happened.