fix: batch size of Discord alerts
Previously if too many operators were affected by an alert and so, the
size of the Discord alert was large, the `alertmanager-discord` package
failed to send these large alerts because of their size.

Here the new feature of the `alertmanager-discord` package that allows
users to configure the size of Discord alerts is adopted. The size of
the Discord alert batch and the number of batches are set in the env
variables of the `alertmanager-discord` package so that now large
Discord alerts should be split into smaller pieces, so now there should
not be errors in sending these pieces to the Discord webhook.