feat: add links with incident time to alerts
In previous versions of alerts, there were links to the "Node Operators"
Grafana dashboard. These links had the end of the time range that
corresponded to the epoch when the incident that triggered the alert
happened. With this timestamp, a user can use the "Alerting Only" panel
in the "Node Operators" dashboard to see the details about validators
affected by this incident.

With a new extended time range in alert links, the end of this time
range no longer points to the epoch that caused the incident, and the
"Alerting Only" panel becomes useless when a user opens the "Node
Operators" dashboard from the links in alerts.

This PR adds links of old format with timestamps that correspond to the
epoch that caused the incident, so users can again use the "Alerting
Only" panel from the alert links.