fix: bug in withdrawal amount calculation
If there is a large amount of withdrawals in the network, it is possible
that in different slots of the same epoch, there might be withdrawals
for the same validator. Previously, it was assumed that in a specific
epoch, there might be only one withdrawal for a specific validator. This
assumption led to a bug.

Now, if there are multiple withdrawals for the same validator in the
same epoch, the amount of withdrawn balance for all these slots is
summed up, so the calculation of the validator balance changes should be
correct now.