fix: broken e2e tests in new version of undici
After the upgrade the undici lib to the latest 6.x version, the e2e
tests get broken. Running tests call the `unref()` function in the
internal components of the undici lib that exists only in the "node"
mode of the test runner and doesn't exist in the
"jest-environment-jsdom" which we are currently using.

It turns out that after upgrading the Jest framework to the more recent
version, the "node" mode of the test runner again starts working
normally. So, in this commit, we get rid of the "jest-environment-jsdom"
mode and remove now unnecessary polyfills. Also, now unnecessary
`jest-environment-jsdom` package has been removed.